[Fix] fix prototype pollution vuln
See https://www.huntr.dev/bounties/1-npm-doc-path/

Note that the link describes pollution via `.prototype` and
`.constructor`, neither of which are actually "prototype pollution".
Mutating a shared object is *supposed* to be visible across the
environment. However, `__proto__` is the one that can be unintentional
and cause these kind of issues.

Note: in node 0.10, `obj['__proto__']` incorrectly does `obj.__proto__`,
which is different according to the spec, so
Object.getOwnPropertyDescriptor is needed.